// src/blogs/types/PageCreate.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

import { BlogInput } from "./../../types/globalTypes";

// ====================================================
// GraphQL mutation operation: PageCreate
// ====================================================

export interface PageCreate_pageCreate_errors {
  __typename: "Error";
  field: string | null;
  message: string | null;
}

export interface PageCreate_pageCreate_page {
  __typename: "Page";
  id: string;
  title: string;
  slug: string;
  isPublished: boolean;
  contentJson: any;
  seoTitle: string | null;
  seoDescription: string | null;
  publicationDate: any | null;
}

export interface PageCreate_pageCreate {
  __typename: "PageCreate";
  errors: PageCreate_pageCreate_errors[] | null;
  blog: PageCreate_pageCreate_page | null;
}

export interface PageCreate {
  blogCreate: PageCreate_pageCreate | null;
}

export interface PageCreateVariables {
  input: BlogInput;
}